requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
asyncssh==2.22.0
attrs==25.4.0
av==16.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.3
csvw==3.7.0
ctranslate2==4.6.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
et_xmlfile==2.0.0
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
invoke==2.2.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro==0.9.4
kubernetes==35.0.0
langchain==1.2.7
langchain-anthropic==1.3.1
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
language-tags==1.2.0
limits==5.6.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
misaki==0.9.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
num2words==0.5.14
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==2.15.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
paramiko==4.0.0
passlib==1.7.4
phonemizer-fork==3.3.2
pillow==12.1.0
posthog==5.4.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.2
pyparsing==3.3.2
PyPDF2==3.0.1
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rdflib==7.5.0
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
segments==2.3.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.5.0
sniffio==1.3.1
soundfile==0.13.1
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
sse-starlette==3.2.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
thinc==8.3.10
tokenizers==0.22.2
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
tqdm==4.67.1
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
